Add Relationships_resizer.txt (#17)
Add the _resizer.txt to the Relationships text. The font will auto resize and it will be able to fit up to 29 characters on the Text Bubbles. Any text longer than 29 characters in lenght will get cut off, and making the font size any lower than 12 is not ideal, the text will look really small.