Registration tutorial translation (#16)
Translates the tutorial on how to register characters that you're shown when making a new save file. The wording feels a bit dry in my opinions so if someone wants to edit it to sound better feel free.
Also fixed a minor regex issue with a group being captured when it shouldn't.